Refactor node main into build node and runner
This will let us build a separate binary for simulator.
